feat: env CAN_DELETE_FILE

+++ src/modules/task/services/task.service.ts
@@ -96,22 +96,24 @@ export class TaskService {
             .where('taskId IN (:...taskIds) and status != :status', { taskIds, status: 1 })
             .execute();
 
-        // 删除文件
-        const whereParams = { isDelete: 1, taskIds };
-        const whereSql = `isDelete = :isDelete and taskId IN (:...taskIds)`;
-        const [data] = await this.taskRepository
-            .createQueryBuilder()
-            .where(whereSql, whereParams)
-            .printSql()
-            .getManyAndCount();
-        data?.filter((task) => !!task?.reportPath)?.forEach((task) => {
-            const filePath = `./static/${task?.reportPath?.replace('/report/', '')}`;
-            try {
-                fs.unlinkSync(filePath);
-            } catch (_error) {
-                console.log(`taskId: ${task.taskId}, 报告文件删除失败，${filePath}`);
-            }
-        });
+        // 删除文件 CAN_DELETE_FILE 值为 yes 时允许删除，本地运行运行删除
+        if (process.env.CAN_DELETE_FILE === 'yes' || process.env.NODE_ENV !== 'production') {
+            const whereParams = { isDelete: 1, taskIds };
+            const whereSql = `isDelete = :isDelete and taskId IN (:...taskIds)`;
+            const [data] = await this.taskRepository
+                .createQueryBuilder()
+                .where(whereSql, whereParams)
+                .printSql()
+                .getManyAndCount();
+            data?.filter((task) => !!task?.reportPath)?.forEach((task) => {
+                const filePath = `./static/${task?.reportPath?.replace('/report/', '')}`;
+                try {
+                    fs.unlinkSync(filePath);
+                } catch (_error) {
+                    console.log(`taskId: ${task.taskId}, 检测报告文件删除失败，${filePath}`);
+                }
+            });
+        }
 
         // 批量删除任务时，把关联的性能数据也删除
         await this.performanceRepository
